+++ R/format-number.r
@@ -307,3 +307,4 @@ number_si <- function(x, prefix = "", suffix = "", ...) {
     ...
   )
 }
+